+++ lib/src/wallet/v3r2.dart
@@ -1,4 +1,3 @@
-import 'dart:convert';
 import 'dart:typed_data';
 
 import '../dataformat/address/api.dart' show InternalAddress, contractAddress;
@@ -52,10 +51,10 @@ class WalletContractV3R2 implements Contract {
 
   WalletContractV3R2(this.publicKey, [this.workChain = 0, int? walletId]) {
     this.walletId = walletId ?? 698983191 + workChain;
-    final decoded = base64.decode(
-        'te6cckEBAQEAcQAA3v8AIN0gggFMl7ohggEznLqxn3Gw7UTQ0x/THzHXC//jBOCk8mCDCNcYINMf0x/TH/gjE7vyY+1E0NMf0x/T/9FRMrryoVFEuvKiBPkBVBBV+RDyo/gAkyDXSpbTB9QC+wDo0QGkyMsfyx/L/8ntVBC9ba0=');
+    final encodedCell =
+        'te6cckEBAQEAcQAA3v8AIN0gggFMl7ohggEznLqxn3Gw7UTQ0x/THzHXC//jBOCk8mCDCNcYINMf0x/TH/gjE7vyY+1E0NMf0x/T/9FRMrryoVFEuvKiBPkBVBBV+RDyo/gAkyDXSpbTB9QC+wDo0QGkyMsfyx/L/8ntVBC9ba0=';
 
-    var code = Cell.fromBoc(Uint8List.fromList(decoded))[0];
+    var code = Cell.fromBocBase64(encodedCell);
     var data = beginCell()
         .storeUint(BigInt.zero, 32) // Seqno: 0
         .storeUint(BigInt.from(this.walletId), 32)